+++ packages/css-calc/src/calculation.ts
@@ -1,5 +1,6 @@
 import type { Operation } from './operation/operation';
 import type { TokenNode } from '@csstools/css-parser-algorithms';
+import type { conversionOptions } from './options';
 import { isTokenNode } from '@csstools/css-parser-algorithms';
 
 export type Calculation = {
@@ -11,7 +12,7 @@ export function isCalculation(x: unknown): x is Calculation {
 	return !!(x) && (typeof x === 'object') && ('inputs' in x) && Array.isArray(x.inputs) && ('operation' in x);
 }
 
-export function solve(calculation: Calculation | -1): TokenNode | -1 {
+export function solve(calculation: Calculation | -1, options: conversionOptions): TokenNode | -1 {
 	if (calculation === -1) {
 		return -1;
 	}
@@ -24,7 +25,7 @@ export function solve(calculation: Calculation | -1): TokenNode | -1 {
 			continue;
 		}
 
-		const result = solve(input);
+		const result = solve(input, options);
 		if (result === -1) {
 			return -1;
 		}
@@ -32,5 +33,5 @@ export function solve(calculation: Calculation | -1): TokenNode | -1 {
 		inputs.push(result);
 	}
 
-	return calculation.operation(inputs);
+	return calculation.operation(inputs, options);
 }