+++ packages/css-calc/src/operation/operation.ts
@@ -1,3 +1,4 @@
 import type { TokenNode } from '@csstools/css-parser-algorithms';
+import type { conversionOptions } from '../options';
 
-export type Operation = (inputs: Array<TokenNode>) => TokenNode | -1
+export type Operation = (inputs: Array<TokenNode>, options: conversionOptions) => TokenNode | -1